+++ changelog.md
@@ -7,6 +7,7 @@
 - 14:30 - Backend: fixed issue when downloaded video file wasn't deleted with the model instance.
 - 17:15 - Backend: added Download model serializer.
 - 18:45 - Frontend: modified Bookmark editor modal to show Downloads-related stuff.
+- 19:00 - Backend: created endpoint to start content download.
 
 ## 21.04.2023, Fri